Issue 1562 - Share link problem - solved
To solve the issue 1562, we analyzed the situation using the UI and then
analyzed the code. We found that sometimes the ArrayList returned by
‘getData’ didn't have only one element, but more, i.e. an “User” link
and a “Public” link. Then we analyzed the code of the
‘RemoteOperationResult’ object and found that it had a method called
'setData'. After this, we identified with the find capability of  the
Android Studio all the locations where ‘setData’ was used and found that
only, ‘ShareToRemoteOperationResultParser‘ and
‘ReadRemoteFolderOperation’ could create an ArrayList with more than one
element. After analyzing what these two classes added to the ArrayList,
before making ‘setData’, we understood that the problem was not in
‘ReadRemoteFolderOperation’. So we verified if there were different
types of links and found in ‘ShareType’ that that was true, being really
dangerous to only accept “Public” links on
‘ShareToRemoteOperationResultParser‘ (it really would give problems),
so we decided to go with the correction presented in the issue with some
adjustments.