Separated the changelog information out into a dedicated file
There is nothing wrong with having the changelog information in with the
rest of the manual.
But, in time, the changelog will become white noise, distracting the
user from the specifics of learning how to use the application.
Given that, I'm separating it out into a separate file.
The information is still there, but suitably structured.